[TS-237] Refactor: mypage component directory change
