+++ src/components/MyPage/Button/Button.tsx
@@ -1,5 +1,5 @@
 import RightArrowIcon from "@/assets/icon/ic-right-arrow.svg?react";
-import { layoutStyle, textStyle, versionTextStyle } from "@/components/MyPage/Button.style";
+import { layoutStyle, textStyle, versionTextStyle } from "@/components/MyPage/Button/Button.style";
 
 interface buttonProps {
 	icon?: JSX.Element;
