+++ src/pages/MyPage.tsx
@@ -1,7 +1,13 @@
 import Gnb from "@/components/common/Gnb/Gnb";
+import Landing from "@/components/MyPage/Landing/Landing";
 
 const MyPage = () => {
-	return <Gnb />;
+	return (
+		<>
+			<Landing />
+			<Gnb />
+		</>
+	);
 };
 
 export default MyPage;
